implemented get_score
added a hand (list of card objects) variable to the player class
modified hit function to add a card to the the player's hand (still needs to detect win/loss)
implemented get_score to return a list of possible scores. If the player has two aces, only one can be counted as 11 at time or else the user will bust no matter what so the max amount of score possibilities is 2.