Completely revamped dealer_turn()
Made it so dealer_turn takes into account multiple players.
dealer_turn uses custom return messages.
Game prints return value of dealer_turn as winning message.
Winning message is printed to terminal, and game is restarted. This is planned to be the message that is parsed for the final game screen. (victory message, losing message)